java exercice

+++ src/bienvenue.java
@@ -63,18 +63,18 @@ public class bienvenue {
             }
         input.close();
 */
-
+/*
         Scanner input = new Scanner(System.in);
         System.out.println("Choix du type operation 1-Addition 2-soustraction 3-multiplication 4-division");
-        int number = input.nextInt();
+        double number = input.nextDouble();
         if (number == 1) {
             System.out.println("operation choisi 1-Addition");
 
             System.out.println("entrer la premiere valeur");
-            int valeur1 = input.nextInt();
+            double valeur1 = input.nextDouble();
 
             System.out.println("entrer la deuxieme valeur");
-            int valeur2 = input.nextInt();
+            double valeur2 = input.nextDouble();
 
             System.out.println (valeur1 + valeur2) ;
         }
@@ -82,10 +82,10 @@ public class bienvenue {
             System.out.println("operation 2-soustraction");
 
             System.out.println("entrer la premiere valeur");
-            int valeur1 = input.nextInt();
+            double valeur1 = input.nextDouble();
 
             System.out.println("entrer la deuxieme valeur");
-            int valeur2 = input.nextInt();
+            double valeur2 = input.nextDouble();
 
             System.out.println (valeur1 - valeur2) ;
         }
@@ -93,10 +93,10 @@ public class bienvenue {
             System.out.println("operation choisi 3-multiplication");
 
             System.out.println("entrer la premiere valeur");
-            int valeur1 = input.nextInt();
+            double valeur1 = input.nextDouble();
 
             System.out.println("entrer la deuxieme valeur");
-            int valeur2 = input.nextInt();
+            double valeur2 = input.nextDouble();
 
             System.out.println (valeur1 * valeur2) ;
         }
@@ -104,14 +104,16 @@ public class bienvenue {
             System.out.println("operation choisi 4-division");
 
             System.out.println("entrer la premiere valeur");
-            int valeur1 = input.nextInt();
+            double valeur1 = input.nextDouble();
 
             System.out.println("entrer la deuxieme valeur");
-            int valeur2 = input.nextInt();
+            double valeur2 = input.nextDouble();
 
             System.out.println (valeur1 / valeur2) ;
 
-        }
+}
+*/
+
